refactor: Migrate things into proper places
