+++ planets-fact-site/app/components/index.ts
@@ -1,0 +1,2 @@
+export { default as CardInformation } from "./card-information";
+export { default as Header } from "./header";
